move all ARCOERA5 code in one module
I renamed the model to a more explicict arcoera5 as ERA5 is the only provided
dataset. Consequently, the global variables used for validation and name
mapping were also renamed as they are only needed for ERA5, so this prefix can
be dropped. I made the variables private as I don't think this implementation
aspect should be exposed to the user, when they load this single module.
