+++ packages/data/src/pyearthtools/data/download/arcoera5.py
@@ -13,11 +13,19 @@
 # limitations under the License.
 
 
-"""
-Translation and default values for `arco.ERA5`
-"""
+from __future__ import annotations
 
-ERA5_LEVELS = [
+import xarray as xr
+
+import pyearthtools.data
+from pyearthtools.data.time import Petdt
+
+from pyearthtools.data.indexes import AdvancedTimeDataIndex, decorators
+from pyearthtools.data.transforms.transform import Transform, TransformCollection
+
+
+# valid ARCO-ERA5 level values
+_VALID_LEVELS = [
     None,
     1,
     2,
@@ -58,7 +66,9 @@ ERA5_LEVELS = [
     1000,
 ]
 
-ERA5_NAME_CHANGE = {
+
+# mapping from long variable names to short variable names
+_LONGNAME_MAPPING = {
     "100m_u_component_of_wind": "u100",
     "100m_v_component_of_wind": "v100",
     "10m_u_component_of_neutral_wind": "u10n",
@@ -334,8 +344,103 @@ ERA5_NAME_CHANGE = {
     "zero_degree_level": "deg0l",
 }
 
-# inverse mapping from short names to full names
-ERA5_NAME_CHANGE_INV = {val: key for key, val in ERA5_NAME_CHANGE.items()}
+# mapping from short variable names to long variable names
+_SHORTNAME_MAPPING = {val: key for key, val in _LONGNAME_MAPPING.items()}
 
-# all variables, short and full names
-ERA5_VARIABLES = list(ERA5_NAME_CHANGE) + list(ERA5_NAME_CHANGE_INV)
+# all valid variables, short and long names
+_VALID_VARIABLES = list(_LONGNAME_MAPPING) + list(_SHORTNAME_MAPPING)
+
+
+def open_arco(variables, level=None, chunks="auto", **kwargs):
+    """Open Analysis-Ready Cloud Optimized ERA5 archive from Google Cloud Platform"""
+
+    # skip parsing unused variables, this can make loading much faster
+    drop_variables = [var for var in _LONGNAME_MAPPING if var not in set(variables)]
+
+    ds = xr.open_zarr(
+        "gs://gcp-public-data-arco-era5/ar/full_37-1h-0p25deg-chunk-1.zarr-v3",
+        chunks=chunks,
+        storage_options=dict(token="anon"),
+        drop_variables=drop_variables,
+        **kwargs,
+    )
+
+    if level is not None:
+        ds = pyearthtools.data.transform.coordinates.Select(level=level, ignore_missing=True)(ds)
+
+    return ds
+
+
+class ARCOERA5(AdvancedTimeDataIndex):
+    """
+    Analysis-Ready, Cloud Optimized ERA5
+
+    https://github.com/google-research/arco-era5
+
+    Carver, Robert W, and Merose, Alex. (2023):
+    ARCO-ERA5: An Analysis-Ready Cloud-Optimized Reanalysis Dataset.
+    22nd Conf. on AI for Env. Science, Denver, CO, Amer. Meteo. Soc, 4A.1,
+    https://ams.confex.com/ams/103ANNUAL/meetingapp.cgi/Paper/415842
+    """
+
+    _desc_ = {
+        "singleline": "Analysis-Ready, Cloud Optimized ERA5",
+        "link": "https://github.com/google-research/arco-era5",
+    }
+
+    @decorators.alias_arguments(
+        variables=["variable"],
+        level=["levels", "level_value"],
+    )
+    @decorators.check_arguments(
+        variables=_VALID_VARIABLES,
+        level=_VALID_LEVELS,
+    )
+    @decorators.variable_modifications("variables")
+    def __init__(
+        self,
+        variables: str | list[str],
+        level: int | list[int] | None = None,
+        transforms: Transform | TransformCollection | None = None,
+        **kwargs,
+    ):
+        """
+        Analysis-Ready, Cloud Optimized ERA5 integrated within `pyearthtools`.
+
+        Allows for access to a cloud ERA5 archive.
+
+        Args:
+            variables (str | list[str]):
+                Variables to retrieve, can be either short_name or long_name
+            level (int | list[int] | None, optional):
+                Pressure levels to select. Defaults to None.
+            transforms (Transform | TransformCollection | None, optional):
+                Transforms to apply to dataset. Defaults to None.
+        """
+        super().__init__(transforms or TransformCollection(), data_interval="1 hour")
+        self.record_initialisation()
+
+        if not isinstance(variables, list):
+            variables = [variables]
+
+        # convert variable name if found in short name mapping
+        variables = [_SHORTNAME_MAPPING.get(var, var) for var in variables]
+
+        self.variables = variables
+        self.level = level
+
+        self._kwargs = kwargs
+        self._ds = open_arco(variables, level, **kwargs)
+
+    @property
+    def dataset(self) -> xr.Dataset:
+        """Get full dataset for this obj"""
+        return self._ds
+
+    def get(self, time: str):
+        """Get timestep from dataset"""
+        return self._ds.sel(time=Petdt(time).datetime64())
+
+    @classmethod
+    def sample(cls):
+        return ARCOERA5("2m_temperature")
